Refactor Redis client package and update import paths
The commit refactors the Redis client package by renaming it to Resp client and
updating the import paths accordingly. The instrumented_redis_client package is also
renamed to instrumented_resp_client. The changes include renaming the package, struct,
and function names, as well as updating the import statements in affected files. This
refactoring improves code organization and makes it more consistent with the package's
functionality.