feat: Add `SendValue()` and related callback.
A new feature has been added to enhance the logging capabilities of our application. Now, before sending a value, an info log will be generated with the formatted string of the value. This is achieved by introducing a new callback function `BeforeSendValue` in both `RespConnectionCallbacks` and `defaultCallbacks`. A corresponding method `SendValue` has also been implemented in `RespConnection`, which calls this callback if it's not nil before sending bytes. This should provide more visibility into what data is being sent from our server.