Refactor SendCommandTestCase to use pointer receiver
This commit refactors the SendCommandTestCase struct's Run method to use a pointer receiver instead of a value receiver. This change allows
modifications to the SendCommandTestCase instance within the Run method to persist outside of the method scope.

Additionally, the commit assigns the value variable to the ReceivedResponse field of the SendCommandTestCase struct, ensuring that the
received response is stored for further use.

Lastly, the commit updates the commandTestCase variable in testReplReplicationID to be a pointer to SendCommandTestCase to match the changes
made in the SendCommandTestCase struct.